move auth files to feature/auth
