+++ src/common/data/serverActions/getUserData.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { createClient } from '@/common/services/auth/supabase/server'
+import { createClient } from '@/features/auth/supabase/server'
 import { User } from '@supabase/supabase-js'
 
 interface GetUserDataActionProps {
