+++ src/common/utils/auth/getUserOrServerRedirect.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/common/services/auth/supabase/server'
+import { createClient } from '@/features/auth/supabase/server'
 import { redirect } from 'next/navigation'
 
 interface Params {
