+++ src/features/jobTracker/data/api/jobApplications.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/common/services/auth/supabase/server'
+import { createClient } from '@/features/auth/supabase/server'
 
 const tableName = 'jobApplications'
